requirements.txt
numpy~=2.0.0
pillow~=10.4.0
Flask~=3.0.3